// src/common/results.enum.ts
export enum ResultsEnum {
  NameMustBeDifferent = 'NAME_MUST_BE_DIFFERENT',
  UsernameMustBeDifferent = 'USERNAME_MUST_BE_DIFFERENT',
  FavoriteSportMustBeDifferent = 'FAVORITE_SPORT_MUST_BE_DIFFERENT',
  FavoriteTimeMustBeDifferent = 'FAVORITE_TIME_MUST_BE_DIFFERENT',
  RoleMustBeDifferent = 'ROLE_MUST_BE_DIFFERENT',
  AvatarMustBeDifferent = 'AVATAR_MUST_BE_DIFFERENT',
  CellphoneMustBeDifferent = 'CELLPHONE_MUST_BE_DIFFERENT',
  NicknameMustBeDifferent = 'NICKNAME_MUST_BE_DIFFERENT',
  UsernameAlreadyInUse = 'USERNAME_ALREADY_IN_USE',
  InvalidPassword = 'INVALID_PASSWORD',
  NewPasswordMustBeDifferent = 'NEW_PASSWORD_MUST_BE_DIFFERENT',
  InvalidCredentials = 'INVALID_CREDENTIALS',
  InvalidToken = 'INVALID_TOKEN',
  TokenExpired = 'TOKEN_EXPIRED',
  PasswordRegex = 'PASSWORD_REGEX',
  NameRegex = 'NAME_REGEX',
  NameSize = 'NAME_SIZE',
  RegistrationSuccessful = 'REGISTRATION_SUCCESSFUL',
  PasswordDoNotMatch = 'PASSWORD_DO_NOT_MATCH',
  InvalidEmail = 'PASSWORD_DO_NOT_MATCH',
  InvalidUserName = 'INVALID_USERNAME',
  PleaseConfirmYourEmail = 'PLEASE_CONFIRM_YOUR_EMAIL',
  YouChangedYourPassword = 'YOU_CHANGED_YOUR_PASSWORD',
  MonthsAgo = 'MONTHS_AGO',
  MonthAgo = 'MONTH_AGO',
  DaysAgo = 'DAYS_AGO',
  DayAgo = 'DAY_AGO',
  HoursAgo = 'HOURS_AGO',
  HourAgo = 'HOUR_AGO',
  Recently = 'RECENTLY',
  LogoutSuccessful = 'LOGOUT_SUCCESSFUL',
  ResetPasswordEmailSent = 'RESET_PASSWORD_EMAIL_SENT',
  PasswordRestSuccessful = 'PASSWORD_RESET_SUCCESSFUL',
  EmailAlreadyConfirmed = 'EMAIL_ALREADY_CONFIRMED',
  EmailAlreadyInUse = 'EMAIL_ALREADY_IN_USE',
  OwnerMustBeArenaType = 'OWNER_MUST_BE_ARENA_TYPE',
  ErrorRegisteringAddress = 'ERROR_REGISTERING_ADDRESS',
  CNPJAlreadyExists = 'CNPJ_ALREADY_EXISTS',
  ErrorRegisteringArena = 'ERROR_REGISTERING_ARENA',
  ArenaDoesNotExists = 'ARENA_DOES_NOT_EXISTS',
  ErrorRegisteringCourt = 'ERROR_REGISTERING_COURT',
  CantCreateRegistrationOnPastDate = 'CANT_CREATE_REGISTRATION_ON_PAST_DATE',
  TimesForThisArenaHaveNotBeenSet = 'TIMES_FOR_THIS_ARENA_HAVE_NOT_BEEN_SET',
  ArenaNotOpenForTheGivenDate = 'ARENA_NOT_OPEN_FOR_THE_GIVEN_DATE',
  YCOCreateReservationOnBusinessHours = 'YCO_CREATE_RESERVATION_ON_BUSINESS_HOURS',
  YCOBookDuringOpeningHours = 'YCO_BOOK_DURING_OPENING_HOURS',
  ThereAlreadyReservationForThisTime = 'THERE_ALREADY_RESERVATION_FOR_THIS_TIME',
  ThereIsAlreadyScheduleForThisDayOfTheWeek = 'BAD_REQ_0001',
}